LUCENE-2257: improve max per-segment term count limit from ~2.1B to ~274B
git-svn-id: https://svn.apache.org/repos/asf/lucene/java/branches/lucene_2_9@909334 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/lucene/index/SegmentTermEnum.java
@@ -108,7 +108,7 @@ final class SegmentTermEnum extends TermEnum implements Cloneable {
     return clone;
   }
 
-  final void seek(long pointer, int p, Term t, TermInfo ti)
+  final void seek(long pointer, long p, Term t, TermInfo ti)
           throws IOException {
     input.seek(pointer);
     position = p;
